Project re-created and rewritten up to that point from the beginning (glitch - SSD). Requests of the type: Get, Post, Put and Delete have been added. Changed the react-bootstrap version to version: 5.2.0 from 6.9.0 due to technical reasons on the replacement machine.

+++ src/main/java/pl/sda/extraproject/bookstore/model/Book.java
@@ -5,7 +5,7 @@ public class Book {
     private String title;
     private String author;
     private String photoUrl;
-    private Integer isbn;
+    private Long isbn;
     private Double price;
 
     public Long getId() {
@@ -40,11 +40,11 @@ public class Book {
         this.photoUrl = photoUrl;
     }
 
-    public Integer getIsbn() {
+    public Long getIsbn() {
         return isbn;
     }
 
-    public void setIsbn(Integer isbn) {
+    public void setIsbn(Long isbn) {
         this.isbn = isbn;
     }
